refactor: meeting -> meetings 변경
